rename dashboard to deployConfigs
