+++ common/actions/index.js
@@ -1,6 +1,6 @@
 export * from './auth'
 export * from './common'
-export * from './dashboard'
+export * from './deployConfigs'
 export * from './deployConfig'
 export * from './inbox'
 export * from './layout'
