+++ common/api/index.js
@@ -1,4 +1,4 @@
 export * from './AuthSvc'
 export * from './InboxSvc'
-export * from './StatisticsSvc'
+export * from './DeployConfigSvc'
 export * from './utils'
