+++ common/containers/index.js
@@ -1,5 +1,5 @@
 export App from './App'
-export Dashboard from './Dashboard'
 export Inbox from './Inbox'
 export Login from './Login'
+export DeployConfigs from './DeployConfigs'
 export DeployConfig from './DeployConfig'
