+++ common/reducers/index.js
@@ -5,14 +5,14 @@ import { layout } from './layout'
 import { inbox } from './inbox'
 import { auth } from './auth'
 import { loginCR } from './login_component_reducer'
-import { dashboard } from './dashboard'
+import { deployConfigs } from './deployConfigs'
 import { deployConfig } from './deployConfig'
 
 export const rootReducer = combineReducers({
   layout,
   inbox,
   auth,
-  dashboard,
+  deployConfigs,
   deployConfig,
   loginCR,
   routing: routerReducer
